+++ apps/client/src/main.tsx
@@ -1,23 +1,14 @@
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
-import { RouterProvider, createRouter } from '@tanstack/react-router';
-import { routeTree } from './routeTree.gen';
 import './global.css';
-
-const router = createRouter({ routeTree });
-
-declare module '@tanstack/react-router' {
-	interface Register {
-		router: typeof router;
-	}
-}
+import App from './App.tsx';
 
 const rootElement = document.getElementById('root')!;
 if (!rootElement.innerHTML) {
 	const root = createRoot(rootElement);
 	root.render(
 		<StrictMode>
-			<RouterProvider router={router} />
+			<App />
 		</StrictMode>
 	);
 }
