+++ apps/client/src/routes/__root.tsx
@@ -1,21 +1,9 @@
-import { createRootRoute, Link, Outlet } from '@tanstack/react-router';
+import { createRootRoute, Outlet } from '@tanstack/react-router';
 import { TanStackRouterDevtools } from '@tanstack/router-devtools';
 
 export const Route = createRootRoute({
 	component: () => (
 		<>
-			<div className="flex gap-2 p-2">
-				<Link to="/" className="[&.active]:font-bold">
-					Harmony
-				</Link>
-				<Link to="/login" className="[&.active]:font-bold">
-					Login
-				</Link>
-				<Link to="/join" className="[&.active]:font-bold">
-					Join
-				</Link>
-			</div>
-			<hr />
 			<Outlet />
 			<TanStackRouterDevtools />
 		</>
