+++ apps/client/src/routes/index.tsx
@@ -1,9 +1,21 @@
-import { createFileRoute } from '@tanstack/react-router';
+import { createFileRoute, Link } from '@tanstack/react-router';
 
-function RouteComponent() {
-	return 'Hello Harmony!';
+function HomePage() {
+	return (
+		<div className="flex gap-2 p-2">
+			<Link to="/" className="[&.active]:font-bold">
+				Harmony
+			</Link>
+			<Link to="/login" className="[&.active]:font-bold">
+				Login
+			</Link>
+			<Link to="/join" className="[&.active]:font-bold">
+				Join
+			</Link>
+		</div>
+	);
 }
 
 export const Route = createFileRoute('/')({
-	component: RouteComponent,
+	component: HomePage,
 });
